Insertion sort: Fixed error in inserting an element into place [Development/Enhancement/Defect] Defect
[Description]
Element to be inserted into its position was being inserted into
array[0] because of an error in determining the swap_position

[Changes]
swap_position is initialised as index itself. When an element is
to be right shifted its shifted from swap_position -1 to
swap_position.

[Test]
Couple of random inputs were given and the algorithm worked fine.